fixing double->float precision loss bug in ThreadSafeRandom.Next, auditing rng usage for entire codebase (#3145)

+++ Source/ACE.Common/ThreadSafeRandom.cs
@@ -10,13 +10,14 @@ namespace ACE.Common
         static readonly ThreadLocal<Random> random = new ThreadLocal<Random>(() => new Random());
 
         /// <summary>
-        /// Returns a random floating-point number between min and max, inclusive
+        /// Returns a random floating-point number that is greater than or equal to 'min', and less than 'max'.
         /// </summary>
-        /// <param name="min">The minimum possible value to return</param>
-        /// <param name="max">The maximum possible value to return</param>
-        public static float Next(float min, float max)
+        /// <param name="min">The value returned will be greater than or equal to 'min'</param>
+        /// <param name="max">The value returned will be less than 'max'</param>
+        public static double Next(float min, float max)
         {
-            return (float)(random.Value.NextDouble() * (max - min) + min);
+            // for ranges other than 1, (max - upper bound) will be scaled by the range
+            return random.Value.NextDouble() * (max - min) + min;
         }
 
         /// <summary>
